+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.2] - 2020-08-04
+* Added docstring to PivotTable.react.js
+* Initial release of dashPivottable R package [#23](https://github.com/plotly/dash-pivottable/pull/23)
 
-## [0.0.1]
+## [0.0.1] - 2020-07-31
 ### Added
 * Up-to-date [Contributing instructions](https://github.com/plotly/dash-pivottable/blob/master/CONTRIBUTING.md)
 * Up-to-date Readme